feat(task-069): decouple silver's schema from the Azure source's physical type (DEC-016 B-minimal)
Claude's half of task-069. Verified by verify-agent in a separate context,
pass on attempt 1. Erik's half — the live Fabric run — remains open and is
NOT claimed anywhere in this commit.

DEC-016 records the fork and Erik's selection. The research changed what the
original two options meant, and added a third:

  - Option B as task-069 framed it turned out to be the SAME three-layer
    migration declined on 2026-08-06 (task-059's documented acceptance) —
    steps 2-5 of the recorded 5-step path, in full. Scope unchanged; only the
    justification was new (a reproducibility-contract failure rather than a
    precision preference). Re-presenting declined work as new work was the
    failure mode worth naming.
  - Option A turned out to be the LARGER documentation change (~3:1), and it
    would have retired the intent data_type_consistency asserts while the
    notebook's own do-not-relabel block forbids the obvious escape.
  - B-minimal is the only formulation genuinely narrower than what was
    declined: one layer moves instead of three.

FIVE code lines:

  1-2. bronze_to_silver casts quantity/unitpriceeur to decimal(18,2) at the
       silver boundary. Silver's schema is now DECLARED rather than inherited
       from an upstream physical type — that inheritance was the actual defect.
  3-4. overwriteSchema on BOTH overwrite branches (full-load and first-load).
       The first-load branch is the one task-069's own description missed.
  5.   silver_to_gold pins gold back to double: (quantity * unitpriceeur)
       .cast("double") and unitpriceeur.cast("double").

The append paths deliberately got NO mergeSchema — adding it would re-hide
exactly the drift this change exists to expose. Three new comments say so in
place so a future maintainer does not "fix" it that way.

The cast order is load-bearing and was verified specifically: the
multiplication happens in decimal and only the RESULT is cast. Measured on
3293.00 x 17.31 (exact 57001.8300) — shipped form 57001.83 (err 0.0);
operand-cast form 57001.829999999994; the pre-change float32-image path
57001.82824134827 (err 1.76e-03).

fact_procurement.tmdl has zero diff lines. That is the whole difference
between B-minimal and B, and it keeps the unverified decimal(37,4) vs Power BI
fixed-decimal(19,4) DirectLake risk off the table entirely.

The quality check was earned, not silenced: type_consistency_checks still
expects `decimal` for both columns, byte-identical. The layer moved to meet the
expectation rather than the expectation being bent to fit.

AC4 closed at SIX sites, not the five DEC-016 enumerated — a sixth identical
stale claim lived in validate_data_type_consistency's own docstring. REAL =
FLOAT(24) -> Spark float; only FLOAT = FLOAT(53) -> double (Microsoft Learn).

AC6 extended: gold_low_confidence_audit.spend_impact stays double only because
unionByName coerces the decimal(38,4) branch against a CAST(NULL AS DOUBLE)
branch — measured, schema-deterministic, so no change needed, but it is luck
rather than a pin. Filed as FR-089.

Local verification CANNOT prove this fix and does not claim to: delta-spark is
not installed and local Spark never invokes a Delta writer, so
DELTA_FAILED_TO_MERGE_FIELDS is structurally unreproducible here. What was
measured locally is the type algebra (real SparkSession, ANSI disabled to match
Fabric), 300/300 tests, and ast.parse on all three notebooks.

+++ fabric/silver_to_gold.Notebook/notebook-content.py
@@ -1665,12 +1665,28 @@ unit_key = F.lower(F.trim(F.col("p.unit")))
 # spend_impact as raw quantity * unitpriceeur (i.e. per_row_unit).
 UNITPRICE_BASIS = "per_row_unit"
 
+# task-069 / DEC-016 (Option B-minimal, 2026-08-12) — GOLD IS PINNED TO `double`.
+# Silver's quantity/unitpriceeur are now decimal(18,2) (declared, not inherited). Left
+# alone, Spark's decimal-multiply rule DECIMAL(p1+p2+1, s1+s2) would make spend_eur
+# decimal(37,4) — a type nobody chose, against Power BI's fixed-decimal decimal(19,4).
+# DirectLake requires the semantic-model type to match the lakehouse type, so that would
+# be an unverified representability risk surfacing in the REPORT layer only, after both
+# notebooks had already succeeded. The `.cast("double")` below keeps
+# fact_procurement.tmdl (unitprice_eur L88-89, spend_eur L99-100, both dataType: double)
+# untouched, so exactly one layer's schema moves: silver.
+#
+# CAST THE RESULT, NEVER THE OPERANDS. The multiplication happens exactly in decimal and
+# only then rounds to the nearest double, so the stored value is ~10^8 closer to the true
+# cent than the old float-image inputs produced. Casting the operands to double first
+# would throw that entire precision benefit away while looking identical in a diff.
+# quantity_base (L1777-1780) already uses exactly this pattern — it is the local
+# convention, not a new concession.
 if UNITPRICE_BASIS == "per_row_unit":
     # Price is EUR per the row's own Unit: multiply the ORIGINAL quantity. No conversion
     # on the price side (converting both sides would double-count), and — crucially —
     # spend does NOT depend on quantity_base, so a non-mass unit (pcs) still gets a real
     # spend even though its quantity_base is NULL.
-    spend_eur_expr = F.col("p.quantity") * F.col("p.unitpriceeur")
+    spend_eur_expr = (F.col("p.quantity") * F.col("p.unitpriceeur")).cast("double")
 elif UNITPRICE_BASIS == "per_kg":
     # Retained for completeness. Price is EUR per kilogram: multiply the kg-normalized
     # quantity. DO NOT use with the current data — it NULLs out every pcs row's spend,
@@ -1829,7 +1845,11 @@ fact_procurement_complete = (
         F.col("supplier_hq_country_key_final").alias("supplier_hq_country_key"),
         F.col("production_country_key_final").alias("production_country_key"),
         F.col("quantity_base"),
-        F.col("p.unitpriceeur").alias("unitprice_eur"),
+        # task-069 / DEC-016: pinned to double. Silver now lands decimal(18,2); without
+        # this cast fact_procurement.unitprice_eur would become decimal(18,2) and
+        # fact_procurement.tmdl L88-89 (dataType: double) would need to change with it.
+        # Pinning here is what keeps the TMDL — and DirectLake — out of this change.
+        F.col("p.unitpriceeur").cast("double").alias("unitprice_eur"),
         F.col("spend_eur"),
         F.col("data_quality_score"),
         F.col("quality_category"),
